requirements.txt
alembic==1.7.4
aniso8601==9.0.1
cachelib==0.4.1
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Dance==5.0.0
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
PyJWT==1.4.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.7
URLObject==2.4.3
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
python-dotenv == 0.21.0
gunicorn == 7.3.1
flask-reuploaded